agentkit/all_in_one_agent/callbacks/before_model_callback.py
import logging
from typing import Optional
from google.adk.agents.callback_context import CallbackContext
from google.adk.models import LlmRequest, LlmResponse

logger = logging.getLogger(__name__)

def before_model_callback(
    callback_context: CallbackContext, llm_request: LlmRequest
) -> Optional[LlmResponse]:
    logger.info("before_model_callback invoked")
    return None
